refactor: format text functions
